ai-agent-hackathon/requirements.txt
# Add requirements here
python-dotenv
google-generativeai
